perms: move perms.go into taprpc
The Lightning Node Connect WASM client requires access to `perms.go` for
RPC endpoint-related permissions. We move it into taprpc so that the
WASM build only needs to depend on the taprpc module, without importing
the entire taproot-assets module.
